claude-3-7/us.anthropic.claude-3-7-sonnet-20250219-v1_0/SecretStorage/Python-FastAPI/temp0.1-openapi-generic/sample0/code/requirements.txt

fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.3
email_validator==2.2.0
python-jose==3.3.0
passlib==1.7.4
pyjwt==2.10.0
requests==2.31.0